Upload media files and store in redux store
Summary:
This diff adds the media files into the redux store. In the redux store they are stored as the following {fbid: description}, where fbid and description are integer and string respectively.

Ignore the rows getting highlighted automatically, there seems to be a bug in the ManagedTable. Talked to Michel and we agreed to replace it.

Reviewed By: mweststrate

Differential Revision: D18708815

fbshipit-source-id: 0aaf1b672c230312203691538091a717d69fb5ad