Tell user when xcode simulator mismatches with xcode-select
Summary:
When you run a simulator with a different version of the xcode version that is currently `xcode-select`ed, it won't work.
This causes problems in Flipper, and also RN and other tools.
Show a helpful warning in this case.

Need to dispatch it explicitly because console.errors only get surfaced in dev mode, not production.

Ideally this will be done in the doctor, but there's work to get it to report device/app specific problems first.

Reviewed By: passy

Differential Revision: D19345199

fbshipit-source-id: 45f95c6d33c81da2299d8d84c9179bebb5bfebe8