Fix refresh plugins not having any effect
Summary:
If a plugin was registered beyond the initial start command, this didn't show up in the Flipper UI. The reason for this is that the client map is updated, but since that object is mutable, components won't detect nor reflect the change.

This diff changes to make sure a new plugin map is created if it is changed.

Reviewed By: passy

Differential Revision: D19345464

fbshipit-source-id: 3aa343e1739caf124a015c88680a94973b87f701