Add isForeground to ping events
Summary:
I want to make ping events fire when in background or foreground.
But doing that now could mess with some stats.
Setting this up so we have the data. If it comes to it, after this has been going a while,
we can add a filter to the existing queries for "isForeground=true" to preserve their behaviour.

Reviewed By: mweststrate

Differential Revision: D19428926

fbshipit-source-id: 7c175e9936dce52a59d17995dae61a82d23687f8