Do not fail if there are entries in errorArray from exportStore
Summary:
Currently `exportStore` function returns the serialized string along with errorArray. errorArray contains an error when there is some issue with fetching metadata. Essentially, it keeps the error in the errorArray and continues with fetching the data for the next plugin. Previously we used to exit our submit form execution when the errorArray had any value. So this diff removes that check and instead logs it on the console and in our logging system.
It will handle the case mentioned [here](https://fb.workplace.com/groups/1430200360634661/permalink/2566831316971554/)

This diff also adds a log when there is failure to upload flipper data.

Reviewed By: mweststrate

Differential Revision: D19723674

fbshipit-source-id: 18bf90461156f67ecc2b4ce4b1eac4aa3485188d