Improve screen capture experience
Summary:
Some improvements on the previous diff that were too much for just comments.

- Remove an unnecessary `touch` that I commented on in the previous diff.
- Ensure that the recording button is not set to "recording" if there's an error during the setup.
- More reliable way of checking the file size on the device before pulling.

Still not done: Show the error more prominently in the UI. Out of scope for this but if we could come up with a design for showing this (perhaps in place of the recording button) when something went wrong, that would be more obvious than having the error caret show up at the top of the screen.

Reviewed By: jknoxville

Differential Revision: D19723627

fbshipit-source-id: 20babcc1482e5a9ac829ff6d6ae7a731a3454fa0