Include positionOnScreen in layout data
Summary:
Context: https://fb.workplace.com/groups/flippersupport/permalink/804152663398727/

This gives flipper the location of every node, in relation to the entire screen. This allows it to visualize the position of each node on it's own diagram, overlayed over a screenshot for example. The use case in mind is after importing flipper data, you'll still be able to see layout nodes overlayed over a screenshot taken during export.

Currently this is only added to Android Views, meaning we don't get it for Litho views or anything on iOS.
This is basically to get something working end-to-end, and if it's all good, then I'll add the other types of views.

Reviewed By: passy

Differential Revision: D19747691

fbshipit-source-id: 8ba3aae6b7685de6faaf55b9628c200802801db4