Fix databases truncated view bug
Summary:
The structure of the page is:
```
<FlexColumn>
  <Toolbar />
  <Content />
  <Toolbar />
</FelxColumn>
```

When Content is a FlexColumn, it doesn't play well with the others and uses all the space, to you don't see the bottom toolbar.
Changing it to a FlexRow fixes this. Also added an inner FlexColumn to make sure the behaviour for things inside Content isn't changed (flex-direction needs to be column in there).

Reviewed By: nikoant

Differential Revision: D19766120

fbshipit-source-id: 2872561bbc954db0d4dc3d7040f60c3eec6e3ff2