Expand deepLinkPayload format for Network tab
Summary: In order to select IDs in a different manner, I am introducing functionality to the Network plugin to allow it to accept more than just IDs that it has assigned, and add the ability to select this plugin with a default search term. I hope to use this to allow for selecting rows by request name.

Reviewed By: mweststrate

Differential Revision: D19702494

fbshipit-source-id: 58ea397fa30eb7d80e9640a4c92905de22bb27d4