Introduce way to clear SearchableTable search term on mount
Summary:
In order to link to this plugin multiple times, we want to be able to clear the existing search term on entry.

This change adds a check to the `componentDidMount` section of `Searchable` to see if we want to update the existing search term from the saved state, or use the one inherited from props.

We call it from the Network tab depending on if an explicit default search term has been set, so normal navigation to and from the plugin still preserves the default behavior -- only when we enter from a `selectPlugin` call does this overwrite the existing term.

Reviewed By: mweststrate

Differential Revision: D19794860

fbshipit-source-id: 4a2b1221acc0732ffc321a16a5eee8348da019f5