Prevent extensive logging of failed network requests
Summary:
If the browser makes a network request, and it fails, this is _always_ logged to the console, despite any error handling being in place. (see https://stackoverflow.com/questions/43012334/silence-neterr-connection-refused)

This diffs doesn't use the browser network stack, but the node network stack instead, which doesn't suffer from the same fate.

Reviewed By: passy

Differential Revision: D19995684

fbshipit-source-id: 4ffc12b820620c5310c140c1a3af63e5d2053a50