Set up oclif
Summary:
Provides some really nice helpers and generators for
multi command CLIs, including test helpers and all
sorts of output niceties. This should make it quite
easy for us to add additional sub-commands for
bundling, publishing and keeping docs for all of it.

Heroku maintains this and provides some excellent
docs, too: https://oclif.io/

My only complaint is that it's class-based but that's
effectively only the way to declare new commands and
enforces a set of required/static properties on it.

Reviewed By: nikoant

Differential Revision: D19970293

fbshipit-source-id: 4228e502198c6fd376854a90ed2f01da29e96bc2