Bundle command scaffolding
Summary:
This sets up a basic UI for a bundling command. It checks for the source, creates an output directory if necessary and automatically creates an output name, so it should hopefully be as intuitive as possible.

There are plenty of potential error case that I don't directly handle. That's something we may want to consider in the future.

Reviewed By: nikoant

Differential Revision: D19973889

fbshipit-source-id: 5638ed0e45187c5ad567b9e2559c8befe52c6085