Fix the edge case to export and import broken export
Summary:
This diff fixes the broken export which is being made [here](https://fb.workplace.com/groups/graphqlandroid/permalink/3319822971399628/). In the import there is no client and plugin states. The fact that the validation succeeded on submit button click suggests that the app got disconnected between the click on submit button and the initiation of export, thus exporting the store with no plugin data and client.

I have added a validation after exporting the store and before the trace is uploaded.

Also in the support request details page, we assumed that client can't be null,  relaxed this assumption as the flipper should still open even the above broken export instead of crashing.

Reviewed By: mweststrate

Differential Revision: D20136295

fbshipit-source-id: f589d01f8ff0ec0b23d53ad5099e11ebbd0930cd