Fix the broken target selector
Summary:
The target selector was broken on iOS. The overlay view which has the gesture recognizer, was never added to the window, due to which the gestures were not registered and the feature broke.

This diff adds the overlay view in the window object as the subview and then discards it in its view descriptor as otherwise we will capture all our touches in this view.

This was the most popular bug in our research trip and almost everyone reported it.

Reviewed By: fabiomassimo

Differential Revision: D20225594

fbshipit-source-id: af4041859d7e662152b2575d1eb258dd0c2d990e