Update iOS deps
Summary:
This diff updates the podfile.lock and also updates all the dependencies in the lfs. Actually we don't need all the dependencies to be reuploaded in the lfs, but current script packs everythin in Pods folder and uploads it, I will upload the logic to look for the diff and only upload the changed dependencies.Also we can merge this logic in our release script. I will put this up in later diffs.

Right now the CK folks are stuck due to the release of CK dependency and their OSS job is failing and this should fix the failing oss-ios job on diff D19948738

Reviewed By: Andrey-Mishanin

Differential Revision: D20378076

fbshipit-source-id: 56ed996cd75fba77a15e7264f9e5c27b262f00d2