Catch DataExportError from shareFlipperData
Summary:
There was a bug where if the `shareFlipperData` returned an `error` with a status code of less than 400 then it will be ignored. If the status code was of greater than or equal to 400 then `shareFlipperData` throws an error and it is caught and logged appropriately.

This diff fixes the above issue where if the return type from shareFlipperData is of type DataExportError then it is checked and not ignored. Also this fixes an issue where if DataExportError happens the loader is replaced with an error message.

Basically the bug mentioned here, T55169042. Also this diff logs the metadata of result, so that we can debug the exact error through scuba.

Reviewed By: mweststrate

Differential Revision: D20450159

fbshipit-source-id: 95b1f384886c16fe444ae0cd6f6d9b2251c29005