Fix the bug where flipper tried to fetch meta data from archived device after importing from same device
Summary:
This diff fixes an issue where when one imports an archived device and tries to export again through the same device with which the initial import was made then the export functionality tries to also fetch metadata from the archived device, it waits for the metadata till it timeouts.

Also this diff fixes an issue where if the selected plugin is not present in the `selectedClient` then it will also ignore that plugin from the other clients even if it was supported by those clients.

Unit test cases for all the above cases are added.

Bug:

{F231519408}

Reviewed By: mweststrate

Differential Revision: D20459179

fbshipit-source-id: 4f0d8c40bec875e3cc43cd6aa70061c8b8da7b05