Refactor Plugin Selection sheet and add snapshot test
Summary:
This diff fixes a UI bug in export data sheet where the plugin names are stuck to the left and also the bottom buttons are stuck together with no padding. I have also added snapshot test for the same. In order to write test and reduce the complexity I changed the `ExportDataPluginSheet`'s connect method.

Bug:

{F231521086}

Reviewed By: mweststrate

Differential Revision: D20459692

fbshipit-source-id: 1047d6b38738691d682ad6e4ccec45c05e14cbbe