Show recent changes automatically at startup
Summary:
This shows a changelog as popup at startup, but only if it wasn't shown before, and only if there are new items in the changelog.

The full changelog can still be accessed through the menu

Changelog: From this release onward we will show important update messages through this dialog.

Reviewed By: passy

Differential Revision: D20492594

fbshipit-source-id: 4663979c8781b468430b9f8b628c4f506578b461