Use http://localhost for origin instead of https
Summary:
Follow up of D20454562

We now set an origin header, but use https instead of http, which is currently not supported by the public RN version (for the internal RN version support was added in D20526486).

However, not promoting ourselves as running on a secured connection seems like the more trustworthy thing to do, and better reflects reality.

Reviewed By: nikoant

Differential Revision: D20534363

fbshipit-source-id: e26fabd77013920e4c58fdbf552bb909f0499cae