Hide timeline by default
Summary:
See the linked task. The QPL can literally freeze Flipper. Initial testing showed that the cause is primary the timeline view.

It was already suggested in the past to hide the timeline by default, as it is not that useful. So did.

Will try to unbreak the timeline in the future, but this at least unblocks the linked task.

Reviewed By: jknoxville

Differential Revision: D20538421

fbshipit-source-id: 8f5cd18c89322e94da5ab055e1913b6434eacf8a