Update podfile for Cocoapods 1.9 (#892)
Summary:
Updates the instructions to integrate iOS pods for the case with the usage of `use_frameworks!` for cocoapods 1.9. Also changed our podfile for Tutorial app and add travis test for the same, so that in future if there are any further changes with the API in new releases of cocoapods, we can catch it.
## Changelog

- Added travis test for iOS Tutorial app
- Updated the getting started guide for iOS
Pull Request resolved: https://github.com/facebook/flipper/pull/892

Test Plan: Travis test passed and also tested that tutorial app builds with cocoapods 1.9

Reviewed By: mweststrate

Differential Revision: D20587989

Pulled By: priteshrnandgaonkar

fbshipit-source-id: b297e6b1298a20a9279ad6414e9d099917876fb1